+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+# Version 3.4.3 (20.11.2018)
+
+## Features
+
+- **chore: Update to GAME_MASTER v1542581210773** 
+
+## Contributors
+
+- [BrunnerLivio](https://github.com/BrunnerLivio/)
+
+# Version 3.4.2 (02.11.2018)
+
+## Features
+
+- **Fix(output)** Normal forms from `form` map
+
+## Contributors
+
+- [BrunnerLivio](https://github.com/BrunnerLivio/)
+- [bytelabsco](https://github.com/bytelabsco)
+
+# Version 3.4.0 (31.10.2018)
+
+## Features
+
+- **Update(GAME_MASTER)** v1542581210773
+- **Add(output)** Pokemon forms attribute
+- **Fix(output)** Incense got removed from the GAME_MASTER.json
+
+## Contributors
+
+- [BrunnerLivio](https://github.com/BrunnerLivio/)
+- [bytelabsco](https://github.com/bytelabsco)
+
+
 # Version 3.3.1 (27.09.2018)
 
 ## Features